🚧 JAVA-3298 Make Factory Inner Class of Resources Collections
