--- src/main/java/com/contrastsecurity/sdk/scan/CodeArtifactsFactory.java
@@ -1,11 +1,0 @@
-package com.contrastsecurity.sdk.scan;
-
-/** Factory for {@link CodeArtifacts} */
-interface CodeArtifactsFactory {
-
-  /**
-   * @param projectId ID of the project in which to manage code artifacts
-   * @return new {@link CodeArtifacts}
-   */
-  CodeArtifacts create(String projectId);
-}
